[gamemaps] Add script for parsing the Wolfenstein 3-D level file format.
It deals decompressing the so-called Carmack compression and the RLEW.

It reads the map header to determine where the levels are located in the
gamemaps then it reads that file for the actual level information.
After parsing the header it can read and decomress the plane. However that
is as far as I've gotten too so far.