[gamemaps] Add some support for Commander Keen 4.
This builds upon the support for Wolf3D as they both use the same file
formats. Command Keen is tricker as the MAPHEAD file is not external
like Wolf3D but is embeeded in the executable. On top of that the
executable itself is compressed.

I would really like to be able to decode and extract the tile sets.