[breakout] Add support for using game pads.
For an Xbox 360 controller this is mapped to the X and B button rather
than the direction pad as that is analog and I can't be bothered dealing
with it.

The keyboard support isn't working for me at the moment, but I can make
do with the gamepad, specially as it eliminates application focus issues.