[breakout] Enabled restarting the game via the game pad.
This required the introduction of the restart_on_next_tick() which is
needed when you want to trigger a restart from say the input thread as
the derived classes may be doing things that can only be accessed on
a certain thread and not this thread.

Ideally, this should probably be using a threading.Event instead of
a boolean like it is.

On an Xbox 360 controller, the select button maps to the start button.