[hgt] Add fast check for if a hgt is a zip when extracting location.
Without this it can be very slow as it tries to read part of the ZIP
file. I am happy to trust the file extention for this.

That said I am wondering if this should check if the file is a zip and
assume it will have a file extention instead.