[dedupe] Introduce the apply program for applying removals to a graph.
Additionally, set-up the make file to make use of this program so
there is now targets that builds an "_after.png" (as well as a dot file).

The C++ code has been sitting around for about two years. The only change
I made today was to trim the whitespace to increase the chance of a match.