[gnaf] Add Python script for transforming SQL scripts to sqlite format.
It essentially merges a script that sets primary keys and foreign keys with
the script that creates the tables.

It uses the sqlparse library to aachive this. I probably could have gotten
away with regular expressions of the first part of the script (finding
the primary key and foreign key) and wrote a simple parse for applying
the changes.