[digital] Add building blocks for defining digital circuits in Python.
The operations module is for working with it algebraic as expression to
compute.

The digital module is provides a way of defining the circuit such that
its preserved and thus offers a level of abstraction such that it
can evaluate the system (give it inputs and get outputs) but can also
output the representation of the system.

The digital module provides an example of an 8-bit adder and is able
to generate a DOT diagram (GraphViz) of it.