[vectortiles] Went with an approach to abstraction walking over the data.
The interface provide lets the caller handle when a layer is enter
and leaf as well as when a feature is introduced.

The alternative that comes to mind is more simple class that wraps the
concept of a tile with layers and features which provides properties that
provide it in a useful form that isn't the underlying protocol buffers.