[vectortiles] Add simplified style for a tile dataset I'm working on.
This should make developing the glstype module easier to get started with
as its not as complicated as the OSM Bright style.

Having said that I only have very minimal interest in continuing with
the Skia-based renderer from Python, but it would be nice to at least
get the poygon/line styles for the states working.