[pdf_viewer] Fit the page to width or height.
The page is now fit such that the page appears in its expected aspect ratio
and so is scaled so a portrait page (what was tested) will have white-space
either side of the page.

Ideally, a black border would be drawn around the page to seperate the
document/page from the background of the application.