[aec] Set-up static website generator for the results.
This is mainly to get a feel for presenting the results. At the moment it
uses a pre-generated CSV file rather than the source data from the AEC's
media feed.

The idea would be a dynamic version where there is a time slider on the
pages to go back in time.