[git-remote-base-rs] Move interaction with on-disk repo to new own module.
The goal at this stage is if you were to write another git remote such as
an S3-like backed one, then the repo module can be used interact with the
repo on-disk that is being cloned to and fetched into.

In theory, it could be used by the file-backed command handler to handle
both the local and remote repos (as the remote one is still on-disk).